refactor Entity Package 추가 (Riot Entity Package 따로 분리 필요할 것 같아서)
