+++ src/test/java/com/gamemoonchul/application/AppleServiceTest.java
@@ -3,7 +3,7 @@ package com.gamemoonchul.application;
 import com.gamemoonchul.common.exception.ApiException;
 import com.gamemoonchul.config.apple.AppleIDTokenValidator;
 import com.gamemoonchul.config.apple.entities.AppleUserInfo;
-import com.gamemoonchul.domain.MemberEntity;
+import com.gamemoonchul.domain.entity.MemberEntity;
 import com.gamemoonchul.infrastructure.repository.MemberRepository;
 import com.gamemoonchul.infrastructure.web.dto.AppleSignUpRequestDto;
 import org.junit.jupiter.api.BeforeEach;
